requirements.txt
ultralytics==8.3.51
opencv-python-headless
pillow
numpy
piexif
transformers
torch
chromadb
timm
rich>=13.0.0  # For pretty error handling
pymilvus